feat: move internal materials internal
