+++ materials/internals/colors/base-colors.mod.css
@@ -2,7 +2,7 @@
 
 /* THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY.
   This file is created by 'scripts/generate-base-color.js' script. The real
-  colors's values should be updated in '../materials/colors/decisions/base-colors.json';
+  colors's values should be updated in '../materials/internals/colors/decisions/base-colors.json';
   Also, This file is transformed to JSON version in order to list the colors
   in the UIKit. Color groups are indicated by a comment preceeding the group
   of colors. */
