+++ materials/internals/colors/colors-for-story.mod.css
@@ -1,6 +1,6 @@
 /* THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY.
   This file is created by 'scripts/generate-colors-for-story.js' script. The real
-  colors's values should be updated in 'materials/colors/decisions/base-colors.json';
+  colors's values should be updated in 'materials/internals/colors/decisions/base-colors.json';
   This file is used to list the colors in the UIKit Color's story purposes only. */
 
 :root {
