+++ materials/internals/colors/colors.story.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import styled from 'react-emotion';
-import { Text } from '../../src';
-import isVariation from '../../src/utils/is-variation';
+import { Text } from '../../../src';
+import isVariation from '../../../src/utils/is-variation';
 import colorGroups from './decisions/base-colors.json';
 import styles from './colors-for-story.mod.css';
