+++ materials/internals/index.css
@@ -1,7 +1,7 @@
-@import './tokens/tokens.mod.css';
+@import './colors/base-colors.mod.css';
 @import './borders.mod.css';
+@import './shadows/base-shadows.mod.css';
 @import './sizes.mod.css';
 @import './spacings.mod.css';
+@import './tokens/tokens.mod.css';
 @import './transitions.mod.css';
-@import './shadows/base-shadows.mod.css';
-@import './colors/base-colors.mod.css';
