+++ materials/internals/shadows/base-shadows.mod.css
@@ -2,7 +2,7 @@
 
 /* THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY.
   This file is created by 'scripts/generate-base-shadows.js' script. The real
-  shadows's values should be updated in '../materials/shadows/decisions/base-shadows.json';
+  shadows's values should be updated in '../materials/internals/shadows/decisions/base-shadows.json';
   Also, This file is transformed to JSON version in order to list the shadows
   in the UIKit. */
