+++ package.json
@@ -19,8 +19,9 @@
     "dist",
     "i18n",
     "images",
-    "materials/*.mod.css",
-    "materials/**/*.mod.css",
+    "materials/materials.css",
+    "materials/materials.js",
+    "materials/media-queries.mod.css",
     "CHANGELOG.md",
     "./README.md",
     "LICENSE",
